backend/requirements.txt
fastapi
uvicorn
openai
whisper
ffmpeg-python
numpy
scikit-learn
python-multipart
python-dotenv
